revamp structure for packaging
