+++ odysseus/odysseus_tree/package/tpu-telemetry/tpu-telemetry.mk
@@ -4,7 +4,7 @@ TPU_TELEMETRY_SITE_METHOD = local
 TPU_TELEMETRY_SETUP_TYPE = setuptools
 
 define TPU_TELEMETRY_INSTALL_INIT_SYSV
-    $(INSTALL) -D -m 0755 $(TPU_TELEMETRY_PKGDIR)/S98tpu-telemetry $(TARGET_DIR)/etc/init.d/S98tpu-telemetry
+    $(INSTALL) -D -m 0755 $(TPU_TELEMETRY_PKGDIR)S98tpu-telemetry $(TARGET_DIR)/etc/init.d/S98tpu-telemetry
 endef
 
 $(eval $(python-package))
