+++ odysseus/odysseus_tree/sources/tpu_telemetry/.gitignore
@@ -1,1 +1,1 @@
-__pycache__
+**/__pycache__
