--- odysseus/odysseus_tree/sources/tpu_telemetry/main.py
@@ -1,73 +1,0 @@
-import asyncio
-# import signal
-import server_data_pb2
-from poll_data import example, can, on_board, halow
-import time
-import asyncio
-from gmqtt import Client as MQTTClient
-
-# initialize connection
-
-client = MQTTClient("tpu-publisher")
-
-STOP = asyncio.Event()
-
-def on_connect(client, flags, rc, properties):
-    print('Connected')
-
-def on_disconnect(client, packet, exc=None):
-    print('Disconnected')
-
-def publish_data(topic, message_data):
-    print(client, topic, message_data)
-
-    # send the data of test
-    client.publish(topic, message_data)
-
-async def initialize():
-
-    host = '127.0.0.1'
-
-    client.on_connect = on_connect # do we need this
-    client.on_disconnect = on_disconnect # do we need this
-
-    # connecting the MQTT broker
-    await client.connect(host, 1883)
-
-async def run():
-    await initialize()
-
-    while True:
-        items = can.fetch_data() + halow.fetch_data() + on_board.fetch_data()
-
-        for file_data in items:
-            print(file_data)
-            topic = file_data[0]
-
-            data = server_data_pb2.ServerData()
-
-            data.unit = file_data[2]
-
-            values = file_data[1]
-            for val in values:
-                data.value.append(val)
-            
-            message_data = data.SerializeToString()
-            
-            publish_data(topic, message_data)
-
-        await asyncio.sleep(1)
-
-
-STOP = asyncio.Event()
-
-if __name__ == '__main__':
-    loop = asyncio.new_event_loop()
-
-    # loop.add_signal_handler(signal.SIGINT, ask_exit)
-    # loop.add_signal_handler(signal.SIGTERM, ask_exit)
-
-    loop.run_until_complete(run())
-
-def ask_exit(*args):
-    STOP.set()
