+++ odysseus/odysseus_tree/sources/tpu_telemetry/pyproject.toml
@@ -12,4 +12,4 @@ dependencies = [
 ]
 
 [project.scripts]
-tpu-telemetry = "main:main"
+tpu-telemetry = "telemetry.main:main"
