--- odysseus/odysseus_tree/sources/tpu_telemetry/server_data.proto
@@ -1,8 +1,0 @@
-syntax = "proto3";
-
-package serverdata.v1;
-
-message ServerData {
-   repeated string value = 1;
-   string unit = 2;
-}
